Relocate functions to correct js file
