+++ src/round.js
@@ -1,40 +1,10 @@
-const { evaluateGuess } = require('./turns');
-
-const giveFeedback = (guess) => {
-  if (guess) {
-    return 'correct!';
-  } else {
-   return 'incorrect!';
+const createRound = (deck) => {
+  return {
+    deck: deck.deck,
+    currentCard: deck.deck[0],
+    turns: 0,
+    incorrectGuesses: [],
   }
-}
-
-const storeIncorrectGuesses = (result, round) => {
-  if(!result){
-    round.incorrectGuesses.push(round.currentCard.id);
-  }
-}
-
-const updateRound = (result, round) => {
-  if (result) {
-    round.deck.shift();
-    } else {
-      let repeat = round.deck.shift();
-      repeat.question = "LET'S TRY AGAIN: " + repeat.question;
-      round.deck.push(repeat);
-    }
-
-  round.turns += 1;
-  round.currentCard = round.deck[0]
-}
-
-const takeTurn = (guess, round) => {
-  let result = evaluateGuess(guess, round.currentCard.correctAnswer);
-  
-  storeIncorrectGuesses(result, round);
-
-  updateRound(result,round);
-
-  return giveFeedback(result);
 }
 
 const calculatePercentCorrect = (round) => {
@@ -51,7 +21,7 @@ const endRound = (round) => {
 }
 
 module.exports = {
-  takeTurn, 
+  createRound,
   calculatePercentCorrect,
   endRound
 }
